% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboost_binnariser.R
\name{apply_xgb_band}
\alias{apply_xgb_band}
\title{apply_xgb_band}
\usage{
apply_xgb_band(x, mapping)
}
\arguments{
\item{x}{A list of column names that identify the explanatory features}

\item{mapping}{A column name that identifies that target variable}
}
\description{
A helper function to apply the mapping table to new data
}
\examples{

}
\keyword{apply}
\keyword{banding}
\keyword{xgboost}
