% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_encoder.R
\name{binary_encoder}
\alias{binary_encoder}
\title{binary_encoder}
\usage{
binary_encoder(df, features, nbits = 32)
}
\arguments{
\item{df}{A data.frame or coercible object that contains the features to be encoded}

\item{features}{A list of names that represent the features to be encoded}

\item{nbits}{How many bits should be encode the information (32 is a good number ;))}
}
\description{
Wrapper function that makes binary encoding of a large number of columns easy
}
\examples{

}
\keyword{binary}
\keyword{encode}
