% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binnarise.R
\name{binnarise}
\alias{binnarise}
\title{binnarise}
\usage{
binnarise(x, w = NULL, nbins = 10, retLabel = FALSE)
}
\arguments{
\item{x}{A vector containing the values to be binned}

\item{w}{A vector that contains the row weights. Defaults to NULL.}

\item{nbins}{How many bins do you wish to create.}
}
\description{
Function that creates equal width bins of a numeric input
}
\examples{

}
\keyword{binnarise}
