% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_double_lift_data.R
\name{create_double_lift_data}
\alias{create_double_lift_data}
\title{create_double_lift_data}
\usage{
create_double_lift_data(modelA, modelB, actual, weight = NULL, nbins = 10)
}
\arguments{
\item{modelA}{A vector that contains the predictions for Model A - preds will be pro-rated}

\item{modelB}{A vector that contains the predictions for Model B - preds will be pro-rated}

\item{actual}{A vector that contains the actuals that the models are to be compared to}

\item{weight}{A vector that contains the row level weights (exposure,claim count). Defaults to NULL}

\item{nbins}{Passed to the binnarise function - how many equally sized bins should the ratio be cut into}
}
\description{
This function creates the requisite data for a double lift chart
}
\examples{

}
\keyword{double}
\keyword{lift}
