% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_interactions.R
\name{create_interactions}
\alias{create_interactions}
\title{create_interactions}
\usage{
create_interactions(df, intList, nlvl = 25, ...)
}
\arguments{
\item{df}{A data.frame that contains the interaction terms to be prepared}

\item{intList}{A list that contains a number of recipes from which to create interactions. Recipes should be character strings delimited by |'s. E.g. "Age|Sex"}

\item{nlvl}{Max number of levels allowed in an interaction term before its binned}

\item{...}{Additional arguments to be passed to the make_interaction}
}
\description{
Wrapper function for make_interaction - assists in the creation of many interaction terms
}
\examples{

}
\keyword{create}
\keyword{interactions}
