% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boruta.R
\name{create_shadow_features}
\alias{create_shadow_features}
\title{create_shadow_features}
\usage{
create_shadow_features(df, features)
}
\arguments{
\item{df}{A data.frame (or coercible object) that contains the features to be shadowed}

\item{features}{A list of feature names to be processed}
}
\description{
Wrapper for .create_shadow to make creating a large number of shadow features easy
}
\examples{

}
\keyword{create}
\keyword{shadow}
