% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{describe}
\usage{
describe(df, cols = NULL)
}
\arguments{
\item{df}{A data.frame (or other coercable object) that contains the columns you wish to describe}

\item{cols}{A list of columns that you wish to describe. If the default value is used all columns will be processed. Defaults to NULL}
}
\description{
This function creates some high level descriptive metrics for columns in a data.frame. Similar to h2o.describe
}
\examples{

}
\keyword{describe}
