% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_encoder.R
\name{.add_noise}
\alias{.add_noise}
\title{.add_noise}
\usage{
.add_noise(series, noise_level)
}
\arguments{
\item{series}{The series to inject noise to}

\item{noise_level}{The level of noise expressed as a proportion of the original series}
}
\description{
This function injects gaussian noise into a numeric series
}
\examples{

}
\keyword{add}
\keyword{noise}
