% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_encoder.R
\name{.binary_encode}
\alias{.binary_encode}
\title{.binary_encode}
\usage{
.binary_encode(x, nbits = 32)
}
\arguments{
\item{x}{A vector that contains the information to be encoded}

\item{nbits}{How many bits should be encode the information (32 is a good number ;))}
}
\description{
Internal function that does binary encoding on a single column. Can give good compression on high cardinality columns.
}
\examples{

}
\keyword{binary}
\keyword{encode}
