% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boruta.R
\name{.create_shadow}
\alias{.create_shadow}
\title{.create_shadow}
\usage{
.create_shadow(x)
}
\arguments{
\item{x}{The vector to be shuffled}
}
\description{
Internal function that creates a randomly shuffled copy of a vector
}
\examples{

}
\keyword{create}
\keyword{shadow}
