% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_encoder.R
\name{.create_smoothing}
\alias{.create_smoothing}
\title{.create_smoothing}
\usage{
.create_smoothing(weights, min_samples_leaf, smoothing)
}
\arguments{
\item{weights}{A vector that contains the weight in each category}

\item{min_samples_leaf}{A scalar that represents the minimum amount of weight deemed to be credible}

\item{smoothing}{A scalar that defines the level of smoothing}
}
\description{
This function smooths a table of category means using an exponential smoothing
}
\examples{

}
\keyword{create}
\keyword{smoothing}
