% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_encoder.R
\name{.label_encode}
\alias{.label_encode}
\title{.label_encode}
\usage{
.label_encode(x)
}
\arguments{
\item{x}{A vector containing the information to be label encoded}
}
\description{
Internal function that label encodes a column
}
\examples{

}
\keyword{encode}
\keyword{label}
