% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_encoder.R
\name{.target_encode}
\alias{.target_encode}
\title{.target_encode}
\usage{
.target_encode(trn_x, tst_x, trn_y, trn_w = NULL, min_samples_leaf = 1,
  smoothing = 1, noise_level = 0)
}
\arguments{
\item{trn_x}{A vector that contains the original x values for the train data}

\item{tst_x}{A vector that contains the x values for the test data}

\item{trn_y}{A vector that contains the train set actuals}

\item{trn_w}{A vector that contains row level weights. Defaults to NULL}

\item{min_samples_leaf}{Whats the smallest amount of weight that's deemed to be credible?}

\item{smoothing}{A parameter that controls the aggressiveness of the smoothing}

\item{noise_level}{How much noise should be injected into the results?}
}
\description{
A worker function that does the target encoding of a column
}
\examples{

}
\keyword{encode}
\keyword{target}
