% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_mdls.R
\name{fit_h2o_mdl}
\alias{fit_h2o_mdl}
\title{fit_h2o_mdl}
\usage{
fit_h2o_mdl(algo = "h2o.glm", x, y, weights_column = NULL, training_frame,
  validation_frame = NULL, ...)
}
\arguments{
\item{algo}{The name of the function used to train the algorithm i.e. "h2o.glm". Defaults to "h2o.glm"}

\item{x}{A list of names indicating the features to be used for training}

\item{y}{A character indicating the target column}

\item{weights_column}{A character indicating the column to be used as row level weights}

\item{training_frame}{An r object pointing to the training data}

\item{validation_frame}{An r object pointing to the validation data}

\item{...}{Additional arguments that will be passed to the underlying training function}
}
\description{
Function to fit any exposed h2o model method. This is a simple wrapper to the h2o api method.
}
\examples{

}
\keyword{fit}
\keyword{h2o}
\keyword{mdl}
