% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boruta.R
\name{get_feat_names}
\alias{get_feat_names}
\title{get_feat_names}
\usage{
get_feat_names(features)
}
\arguments{
\item{features}{A list of feature names to be processed}
}
\description{
Noddy function to get shadow feature names
}
\examples{

}
\keyword{names}
\keyword{shadow}
