% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_mdls.R
\name{get_h2o_predictions}
\alias{get_h2o_predictions}
\title{get_h2o_predictions}
\usage{
get_h2o_predictions(mdl, newdata = NULL, xval = FALSE)
}
\arguments{
\item{mdl}{An h2o model object to predict from}

\item{newdata}{A new data set to make predictions on. Defaults to NULL}

\item{xval}{Do you want to extract the out of fold predictions? Defaults to FALSE. Note: keep_cross_validation_predictions must equal TRUE in the model object}
}
\description{
Function to return predictions from an h2o model to a local r object
}
\examples{

}
\keyword{h2o}
\keyword{predict}
