% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_interactions_untested.R
\name{interaction_search}
\alias{interaction_search}
\title{interaction_search}
\usage{
interaction_search(mdl, features, XGBfiBin = "~/Kaggle/xgbfi/bin/",
  filterExp = NULL, XGBfiParams = list(d = 3, g = -1, t = 100, k = 100, h =
  10), intern = FALSE)
}
\arguments{
\item{mdl}{A trained xgboost model of class xgb.Booster}

\item{features}{A list of the feature names used in the model (ordering must be consistent with the dmatrix used to train the model)}

\item{XGBfiBin}{The file path to the bin of the XGBfi utility}

\item{filterExp}{A character expression that tells the process how to filter the XGBfi results. e.g. Gain > 2000}

\item{XGBfiParams}{A named list that contains the parameters to be passed to the cli of the XGBfi utility.}

\item{intern}{Passed to the system command (internalise console output or not)}
}
\description{
This function runs the XGBfi utility on a trained xgboost model to identify interactions to be included in future model interations.
}
\examples{

}
\keyword{interaction}
\keyword{search}
