% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_encoder.R
\name{label_encoder}
\alias{label_encoder}
\title{label_encoder}
\usage{
label_encoder(df, features)
}
\arguments{
\item{df}{A data.table (or coercible object) that contains the columns to be encoded}

\item{features}{A list of column names to be encoded}
}
\description{
Wrapper for .label_encode to aid in the label encoding of many columns
}
\examples{

}
\keyword{label_encoder}
