% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2o_mdls.R
\name{load_h2o_data}
\alias{load_h2o_data}
\title{load_h2o_data}
\usage{
load_h2o_data(data, key, write_read = NULL)
}
\arguments{
\item{data}{An r data.frame to be transfered into the h2o cluster}

\item{key}{The character key that will be used to identify the h2o.frame in the cluster}

\item{write_read}{A logical allowing the user to specify which method to be used}
}
\description{
Function to load R objects into h2o.Frames - built the write/read trick under certain circumstances.
}
\examples{

}
\keyword{data}
\keyword{h2o}
\keyword{load}
