% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_interactions.R
\name{make_interaction}
\alias{make_interaction}
\title{make_interaction}
\usage{
make_interaction(df, recipe, nlvl, ...)
}
\arguments{
\item{df}{A data.frame that contains the interaction terms to be prepared}

\item{recipe}{A character string containing interaction terms delimited by | e.g. "age|sex".}

\item{nlvl}{Max number of levels allowed in an interaction term before its binned}

\item{...}{Additional arguments to be passed to the prepare interaction function}
}
\description{
Function that creates an interaction term from a recipe
}
\examples{

}
\keyword{interaction}
\keyword{make}
