% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot_encoder.R
\name{one_hot_encoder}
\alias{one_hot_encoder}
\title{One Hot Encoder}
\usage{
one_hot_encoder(df, features)
}
\arguments{
\item{df}{A data.frame (or coercible) object that contains the columns to be encoded}

\item{features}{A list of names that identify the columns to be encoded}
}
\description{
This function one-hot encodes a list of features
}
\keyword{encode}
\keyword{hot}
\keyword{one}
