% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ave_data.R
\name{plot_ave_data}
\alias{plot_ave_data}
\title{plot_ave_data}
\usage{
plot_ave_data(x, y_true, y_pred = NULL, w = NULL, rescale = FALSE)
}
\arguments{
\item{x}{A vector that contains the dimension that you want to aggregate by}

\item{y_true}{A vector that contains the target variable (think OBS in emblem)}

\item{y_pred}{A vector that contains the fitted value from a model (think CA in emblem). Defaults to NULL}

\item{w}{A vector that contains row level weights (e.g. exposure). Defaults to NULL}

\item{rescale}{A logical indicating whether you want to the resulting series scaled about the largest level}
}
\description{
A companion function to create_ave_plot_data. This function calls the aggregator and then makes a nice plot.
}
\examples{

}
\keyword{ave}
\keyword{plot}
