% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantile_data.R
\name{plot_quantile_data}
\alias{plot_quantile_data}
\title{plot_quantile_data}
\usage{
plot_quantile_data(y_pred, y_true, w = NULL, nbins = 10)
}
\arguments{
\item{y_pred}{The predictions that you want to compare the actuals to - these will be pro-rated.}

\item{y_true}{The actuals you want to compare your model against}

\item{w}{A column of row level weights (e.g. exposure, claim count). Defaults to NULL}

\item{nbins}{How many quantiles do you want to create? Defaults to 10}
}
\description{
A companion function to create_quantile_plot_data. It runs the aggregator and then makes a nice plot.
}
\examples{

}
\keyword{plot}
\keyword{quantile}
