% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_example_data.R
\name{prepare_example_data}
\alias{prepare_example_data}
\title{prepare_example_data}
\usage{
prepare_example_data()
}
\description{
Internal function that label encodes a column
}
\examples{

}
\keyword{data}
\keyword{example}
