% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boruta.R
\name{process_importances}
\alias{process_importances}
\title{process_importances}
\usage{
process_importances(importances)
}
\arguments{
\item{importances}{a list of xgboost style importances}
}
\description{
A function to aggregate a list of importances into some useful feature selection information
}
\examples{

}
\keyword{importances}
\keyword{process}
