% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebase_col.R
\name{rebase_col}
\alias{rebase_col}
\title{rebase_col}
\usage{
rebase_col(x, base, w = NULL)
}
\arguments{
\item{x}{The column that you wish to rebase}

\item{base}{The column that you wish to rebase x to}

\item{w}{A vector of row weights. Defaults to NULL}
}
\description{
A function that rebases a column to another column. Accepts row weights.
}
\examples{

}
\keyword{col}
\keyword{rebase}
