% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_colinear_features.R
\name{remove_colinear_features}
\alias{remove_colinear_features}
\title{remove_colinear_features}
\usage{
remove_colinear_features(df, features, threshold = 0.75, ...)
}
\arguments{
\item{df}{A data.frame that contains the columns to be processed}

\item{features}{A character vector containing the colnames to be processed}

\item{threshold}{What level of correlation is too high? Defaults to 0.75}

\item{...}{Additional arguments to be passed to stats::cor}
}
\description{
Function remove colinear features
}
\examples{

}
\keyword{colinear}
\keyword{remove}
