% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_encoder.R
\name{target_encoder}
\alias{target_encoder}
\title{target_encoder}
\usage{
target_encoder(train_df, test_df, y, x, w = NULL, verbose = FALSE, ...)
}
\arguments{
\item{train_df}{A data.frame (or coercible object) that contains the training set columns to be encoded}

\item{test_df}{A data.frame (or coercible object) that contains the test set columns to be encoded}

\item{y}{A string containing the name of the target column}

\item{x}{A string containing the names of the columns to be processed}

\item{w}{A string containing the name of the weight column. Defaults to NULL.}

\item{verbose}{A logical indicating whether to print run time progress messages or not.}

\item{...}{Further arguments to be passed to .target_encode}
}
\description{
A wrapper for .target_encode that makes target encoding a large number of columns easy
}
\examples{

}
\keyword{encode}
\keyword{target}
