% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_xgoost.R
\name{xgb_create_dmatrix}
\alias{xgb_create_dmatrix}
\title{train_xgboost}
\usage{
xgb_create_dmatrix(data, x, y, w = NULL, base_margin = NULL)
}
\arguments{
\item{data}{A data.frame (or coercible object) that contains the training data}

\item{x}{A list of column names that identify the explanatory features}

\item{y}{A column name that identifies that target variable}

\item{w}{A column name that contains the name of the weight column. Defaults to NULL}

\item{w}{A column name that contains the name of the offset column. Defaults to NULL}
}
\description{
A wrapper function that automates the xgboost training process
}
\examples{

}
\keyword{create}
\keyword{dmatrix}
