% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boruta.R
\name{xgboost_boruta}
\alias{xgboost_boruta}
\title{xgboost_boruta}
\usage{
xgboost_boruta(df, x, y, w = NULL, niter = 10, xgbParams)
}
\arguments{
\item{df}{A data.frame (or coercible object) that contains the training data}

\item{x}{A list of feature names to be appraised}

\item{y}{The name of the target column}

\item{w}{The name of the weights column. defaults to NULL.}

\item{niter}{How many xgboost models should we train}

\item{xgbParams}{List of learning parameters to be passed to XGBoost}
}
\description{
Xgboost flavoured boruta implementation for feature selection. WARN: This method is all brute force so training time can be long.
}
\examples{

}
\keyword{boruta}
\keyword{xgboost}
